+++ FacebookClone/FacebookClone.Presentation/EndpointDefinitions/CommentEndpointDefinitions.cs
@@ -1,5 +1,6 @@
 ﻿using FacebookClone.BLL.DTO.Comment.Friendship;
 using FacebookClone.BLL.Services.Abstract;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace FacebookClone.Presentation.EndpointDefinitions
@@ -8,19 +9,19 @@ namespace FacebookClone.Presentation.EndpointDefinitions
     {
         public static void DefineEndpoints(WebApplication app)
         {
-            app.MapGet("/comments", (ICommentService commentService) => commentService.GetAll());
+            app.MapGet("/comments", [Authorize(Policy = "RequireId")] (ICommentService commentService) => commentService.GetAll());
 
-            app.MapPost("/comments", (CommentDTO comment, ICommentService commentService) => commentService.Add(comment));
+            app.MapPost("/comments", [Authorize(Policy = "RequireId")] (CommentDTO comment, ICommentService commentService) => commentService.Add(comment));
 
-            app.MapGet("/comments/{id}", (ICommentService commentService, int id) => commentService.GetById(id));
+            app.MapGet("/comments/{id}", [Authorize(Policy = "RequireId")] (ICommentService commentService, int id) => commentService.GetById(id));
 
-            app.MapGet("/comments/search/{userId}", (ICommentService commentService, int userId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => commentService.GetAllByUserId(userId, pageSize, pageNumber));
+            app.MapGet("/comments/search/{userId}", [Authorize(Policy = "RequireId")] (ICommentService commentService, int userId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => commentService.GetAllByUserId(userId, pageSize, pageNumber));
 
-            app.MapGet("/comments/album/{albumId}", (ICommentService commentService, int albumId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => commentService.GetAllByAlbumId(albumId, pageSize, pageNumber));
+            app.MapGet("/comments/album/{albumId}", [Authorize(Policy = "RequireId")] (ICommentService commentService, int albumId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => commentService.GetAllByAlbumId(albumId, pageSize, pageNumber));
 
-            app.MapDelete("/comments/{id}", (ICommentService commentService, int id) => commentService.Delete(id));
+            app.MapDelete("/comments/{id}", [Authorize(Policy = "RequireId")] (ICommentService commentService, int id) => commentService.Delete(id));
 
-            app.MapPut("/comments", (CommentDTO comment, ICommentService commentService) => commentService.Update(comment));
+            app.MapPut("/comments", [Authorize(Policy = "RequireId")] (CommentDTO comment, ICommentService commentService) => commentService.Update(comment));
         }
     }
 }
